Refactors home page
