+++ features/hero-banner/hero-banner.style.ts
@@ -1,0 +1,32 @@
+import styled from "@emotion/styled";
+
+export const BodyContainer = styled.div`
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  padding-top: 2rem;
+`;
+export const Container = styled.div`
+  width: 64rem;
+  height: 20rem;
+  background: #f2c94d url("/icons/captain-marvel.jpg") no-repeat right top;
+  background-size: 39.375rem;
+  border-radius: 0.625rem;
+`;
+export const TextWrapper = styled.div`
+  display: flex;
+  flex-direction: column;
+  padding: 2rem;
+`;
+export const H1 = styled.h1`
+  font-size: 2.813rem;
+  letter-spacing: 0.1rem;
+  margin-bottom: 0;
+`;
+export const InfoText = styled.p`
+  font-size: 1.125rem;
+  background-color: rgba(255, 255, 255, 0.25);
+  border-radius: 0.5rem;
+  padding: 0.5rem;
+  margin-right: 400px;
+`;
