+++ features/hero-banner/index.ts
@@ -1,0 +1,1 @@
+export { HeroBanner } from "./hero-banner";
