+++ features/index.ts
@@ -1,2 +1,3 @@
 export * from "./page-container";
 export * from "./hero-banner";
+export * from "./search-container";
