+++ features/page-container/index.ts
@@ -1,0 +1,1 @@
+export { PageContainer } from "./page-container";
