+++ features/search-container/index.ts
@@ -1,0 +1,1 @@
+export { SearchContainer } from "./search-container";
