+++ features/search-container/search-container.style.ts
@@ -1,0 +1,14 @@
+import styled from "@emotion/styled";
+
+export const Container = styled.div`
+  display: flex;
+  align-items: center;
+  width: 100%;
+  min-height: 3.5rem;
+  background-color: #343a40;
+`;
+
+export const Title = styled.p`
+  padding-left: 7.5rem;
+  color: white;
+`;
