+++ features/search-container/search-container.tsx
@@ -1,0 +1,10 @@
+import React from "react";
+import { Container, Title } from "./search-container.style";
+
+export function SearchContainer() {
+  return (
+    <Container>
+      <Title>Marvelous</Title>
+    </Container>
+  );
+}
