+++ pages/index.tsx
@@ -1,49 +1,5 @@
 import Head from "next/head";
-import styled from "@emotion/styled";
-import { HeroBanner, PageContainer } from "@features";
-
-const Skyline = styled.div`
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  background-image: url("/icons/sky-line.png"),
-    linear-gradient(
-      to bottom,
-      #163c52 0%,
-      #4f4f47 30%,
-      #c5752d 60%,
-      #b7490f 80%,
-      #2f1107 100%
-    );
-  background-size: contain, auto;
-  background-position: bottom;
-  background-repeat: repeat-x;
-  min-height: 200px;
-`;
-const HeaderName = styled.h1`
-  font-size: 5rem;
-  margin: 0;
-  letter-spacing: 3px;
-  transform: rotate(-5deg) skew(-10deg, 0);
-  text-shadow: -2px -2px yellow, 2px 2px orange, 3px 3px orange;
-
-  @media (max-width: 375px) {
-    font-size: 3rem;
-  }
-`;
-
-const SearchContainer = styled.div`
-  display: flex;
-  align-items: center;
-  width: 100%;
-  min-height: 3.5rem;
-  background-color: #343a40;
-`;
-
-const Title = styled.p`
-  padding-left: 7.5rem;
-  color: white;
-`;
+import { HeroBanner } from "@features";
 
 export default function Home() {
   return (
@@ -53,19 +9,7 @@ export default function Home() {
         <meta name="description" content="Generated by create next app" />
         <meta name="viewport" content="width=device-width, initial-scale=1" />
       </Head>
-      <header>
-        <Skyline>
-          <HeaderName data-cy="h1">MARVELOUS</HeaderName>
-        </Skyline>
-      </header>
-      <div>
-        <SearchContainer>
-          <Title>Marvelous</Title>
-        </SearchContainer>
-        <PageContainer>
-          <HeroBanner />
-        </PageContainer>
-      </div>
+      <HeroBanner />
     </>
   );
 }
